feat: buzzer screen for admin that is not a registered player
fixes: #213

feat: Create and store a host password outside of game data to
authenticate functions that only the host should do.

feat: after joining a game session there is a button to go to the
"/buzzers" page that with user session and host password will be a page
with 2 large button icons that will trigger the buzzer for team 1 and
team 2. This page will also listen on keyboard key presses to activate
the buzzer so users have potential to hook up physical buttons to a
device to trigger the buzzers

fix(frontend): Create a type, enum for defining the incoming events and actions

Start playing with React Contexts for passing down global state

fix(admin): supply hostPassword on send events from admin page